website/pyproject.toml
[tool.poetry]
name = "luleo"
version = "0.1.0"
description = ""
authors = ["Raghuvar Nadig <raghu@luleo.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
firebase-admin = "^6.5.0"
pytz = "^2024.1"
poetry-dotenv = "^0.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
